🐛 [BE] 서버간 통신 polling 타입 long에서 int로 수정

+++ backend/chat/src/main/java/com/ssafy/chat/common/service/AbstractChatAuthService.java
@@ -47,7 +47,7 @@ public abstract class AbstractChatAuthService {
 
         // 인증 결과 대기
         Map<String, Object> authResult = chatAuthResultService.waitForAuthResult(
-                requestId, chatRoomUtils.getAuthTimeoutMs());
+                requestId, (int) chatRoomUtils.getAuthTimeoutMs());
 
         if (authResult == null) {
             log.error("bbatty 서버 인증 응답 타임아웃 - requestId: {}", requestId);
